+++ packages/node/src/indexer/worker/worker.ts
@@ -22,6 +22,7 @@ import { INestApplication } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { registerWorker, getLogger, NestLogger, waitForHeap } from '@subql/node-core';
 import { SpecVersion } from '../dictionary.service';
+import { DynamicDsService } from '../dynamic-ds.service';
 import { IndexerManager } from '../indexer.manager';
 import { WorkerModule } from './worker.module';
 import {
@@ -80,7 +81,15 @@ async function fetchBlock(
 async function processBlock(height: number): Promise<ProcessBlockResponse> {
   assert(workerService, 'Not initialised');
 
-  return workerService.processBlock(height);
+  const res = await workerService.processBlock(height);
+
+  // Clean up the temp ds records for worker thread instance
+  if (res.dynamicDsCreated) {
+    const dynamicDsService = app.get(DynamicDsService);
+    dynamicDsService.deleteTempDsRecords(height);
+  }
+
+  return res;
 }
 
 function syncRuntimeService(