+++ packages/query/src/configure/configure.module.ts
@@ -1,17 +1,32 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import {ConnectionOptions} from 'tls';
 import {DynamicModule, Global, Module} from '@nestjs/common';
-import {Pool} from 'pg';
+import {getFileContent, CONNECTION_SSL_ERROR_REGEX} from '@subql/common';
+import {Pool, PoolConfig} from 'pg';
 import {getLogger} from '../utils/logger';
 import {getYargsOption} from '../yargs';
 import {Config} from './config';
 import {debugPgClient} from './x-postgraphile/debugClient';
 
+async function ensurePool(poolConfig: PoolConfig): Promise<Pool> {
+  const pgPool = new Pool(poolConfig);
+  try {
+    await pgPool.connect();
+  } catch (e) {
+    if (JSON.stringify(e.message).includes(CONNECTION_SSL_ERROR_REGEX)) {
+      poolConfig.ssl = undefined;
+      return ensurePool(poolConfig);
+    }
+  }
+  return pgPool;
+}
+
 @Global()
 @Module({})
 export class ConfigureModule {
-  static register(): DynamicModule {
+  static async register(): Promise<DynamicModule> {
     const {argv: opts} = getYargsOption();
 
     const config = new Config({
@@ -20,7 +35,29 @@ export class ConfigureModule {
       unsafe: opts.unsafe ?? false,
     });
 
-    const pgPool = new Pool({
+    const dbSslOption = () => {
+      const sslConfig: ConnectionOptions = {rejectUnauthorized: false};
+      if (opts['pg-ca']) {
+        try {
+          sslConfig.ca = getFileContent(opts['pg-ca'], 'postgres ca cert');
+          if (opts['pg-key']) {
+            sslConfig.key = getFileContent(opts['pg-key'], 'postgres client key');
+          }
+
+          if (opts['pg-cert']) {
+            sslConfig.cert = getFileContent(opts['pg-cert'], 'postgres client cert');
+          }
+
+          return sslConfig;
+        } catch (e) {
+          getLogger('db config').error(e);
+          throw e;
+        }
+      }
+      return sslConfig;
+    };
+
+    const poolConfig: PoolConfig = {
       user: config.get('DB_USER'),
       password: config.get('DB_PASS'),
       host: config.get('DB_HOST_READ') && !opts.subscription ? config.get('DB_HOST_READ') : config.get('DB_HOST'),
@@ -28,7 +65,11 @@ export class ConfigureModule {
       database: config.get('DB_DATABASE'),
       max: opts['max-connection'],
       statement_timeout: opts['query-timeout'],
-    });
+      ssl: dbSslOption(),
+    };
+
+    const pgPool = await ensurePool(poolConfig);
+
     pgPool.on('error', (err) => {
       // tslint:disable-next-line no-console
       getLogger('db').error('PostgreSQL client generated error: ', err.message);